refactor(jobs): split Storage into focused files, eliminate SQL duplicates
Split storage.go (1509 lines) into 7 files by concern:
- storage.go: struct + constructor (13 lines)
- storage_functions.go: job function CRUD (18 methods)
- storage_function_files.go: file management (3 methods)
- storage_queue.go: job lifecycle (13 methods)
- storage_queries.go: job queries + stats (5 methods)
- storage_workers.go: worker management (8 methods)
- storage_namespaces.go: namespace queries (2 methods)

Eliminate 3 exact SQL duplicates:
- GetJobByIDAdmin -> delegates to GetJob
- ListJobsAdmin -> delegates to ListJobs
- CreateJob already aliased EnqueueJob
